backend/requirements.txt
# Django & REST
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
django-ratelimit>=4.1,<5.0

# Database
psycopg[binary]>=3.1,<4.0

# Async tasks
celery>=5.4,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0

# Email & SMS
twilio>=9.0,<10.0

# File generation
openpyxl>=3.1,<4.0
qrcode[pil]>=7.4,<8.0

# HTTP requests
requests>=2.31,<3.0

# Utilities
python-decouple>=3.8,<4.0
Pillow>=10.0,<11.0
ua-parser>=1.0,<2.0

# Production
gunicorn>=22.0,<23.0
whitenoise>=6.6,<7.0
